refactor(TracePage): extract keyboard shortcut setup into dedicated module
Move keyboard shortcut registration and teardown logic from TracePage's
componentDidMount/componentWillUnmount into a new keyboard-setup.ts
module. This is a pure code movement with no behavior changes.

The new setupTracePageShortcuts() function takes the scroll manager and
callback dependencies as parameters, registers all shortcuts via
mergeShortcuts(), and returns a cleanup function that calls
resetShortcuts() and cancelScroll().

Also moves shortcutConfig, makeShortcutCallbacks, VIEW_CHANGE_BASE and
VIEW_CHANGE_FAST into the new module. These are re-exported from
index.tsx for backward compatibility with existing tests.

Part of the incremental TracePage migration (#3381).
